refactor(lowy): cross-validate — HostSessionLike belongs in kolu-remote-client, not kolu-shared
Lowy cross-validation of hickey's earlier dedup move flagged a
layering inversion: kolu-shared was previously utilities-only ('no
agent-specific concepts'), and placing a high-volatility transport
interface there maximises blast radius — every package transitively
depends on kolu-shared, but only ~3 actually consume HostSessionLike.

The cycle premise hickey was avoiding doesn't exist: kolu-pty can
add kolu-remote-client as a dep with no loop (kolu-remote-client
does not depend on kolu-pty).

Fix:
- Move host-session.ts back to packages/kolu-remote-client/src/
- Re-export HostSessionLike from kolu-remote-client's index
- Delete packages/kolu-remote-client/src/types.ts (was a thin re-export)
- Restore kolu-shared/index.ts to its prior shape
- Add kolu-remote-client to kolu-pty's deps; agent-pty-provider
  imports HostSessionLike from kolu-remote-client
