+++ packages/integrations/pty/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@xterm/addon-serialize": "^0.14.0",
     "@xterm/headless": "^6.0.0",
+    "kolu-remote-client": "workspace:*",
     "kolu-shared": "workspace:*",
     "node-pty": "github:juspay/node-pty#fix/darwin-pty-fd-leak"
   },
